Novos Exs.

+++ Exercícios/61-80/ex-064.py
@@ -1,7 +1,7 @@
 aluno = dict()
 
 aluno["Nome"] = input("Digite o seu nome: ")
-aluno["Média"] = int(input("Digite a sua média: "))
+aluno["Média"] = float(input("Digite a sua média: "))
 aluno["Situação"] = "Aprovado" if (aluno["Média"] >= 9.5) else "Reprovado"
 
 print(aluno)
